feat(wallet-home): wallet_home_screen/view_model.dart 추가 & 기존 wallet_list 화면 및 로직 이동. wallet_list 불필요한 요소 삭제
