+++ lib/providers/view_model/home/wallet_home_view_model.dart
@@ -4,7 +4,10 @@ import 'package:coconut_wallet/enums/network_enums.dart';
 import 'package:coconut_wallet/model/wallet/balance.dart';
 import 'package:coconut_wallet/model/wallet/wallet_list_item_base.dart';
 import 'package:coconut_wallet/providers/connectivity_provider.dart';
+import 'package:coconut_wallet/providers/preference_provider.dart';
+import 'package:coconut_wallet/providers/visibility_provider.dart';
 import 'package:coconut_wallet/providers/wallet_provider.dart';
+import 'package:coconut_wallet/services/app_review_service.dart';
 import 'package:coconut_wallet/utils/vibration_util.dart';
 import 'package:flutter/material.dart';
 
@@ -12,33 +15,52 @@ typedef AnimatedBalanceDataGetter = AnimatedBalanceData Function(int id);
 typedef BalanceGetter = int Function(int id);
 typedef FakeBalanceGetter = int? Function(int id);
 
-class WalletListViewModel extends ChangeNotifier {
+class WalletHomeViewModel extends ChangeNotifier {
+  late final VisibilityProvider _visibilityProvider;
   WalletProvider _walletProvider;
   final Stream<NodeSyncState> _syncNodeStateStream;
+  late final PreferenceProvider _preferenceProvider;
+  late bool _isTermsShortcutVisible;
+  late bool _isBalanceHidden;
+  late final bool _isReviewScreenVisible;
   late final ConnectivityProvider _connectivityProvider;
   late bool? _isNetworkOn;
   Map<int, AnimatedBalanceData> _walletBalance = {};
-
+  Map<int, dynamic> _fakeBalanceMap = {};
+  int? _fakeBalanceTotalAmount;
   bool _isFirstLoaded = false;
   NodeSyncState _nodeSyncState = NodeSyncState.syncing;
   StreamSubscription<NodeSyncState>? _syncNodeStateSubscription;
 
-  WalletListViewModel(
+  WalletHomeViewModel(
     this._walletProvider,
+    this._preferenceProvider,
+    this._visibilityProvider,
     this._connectivityProvider,
     this._syncNodeStateStream,
   ) {
+    _isTermsShortcutVisible = _visibilityProvider.visibleTermsShortcut;
+    _isReviewScreenVisible = AppReviewService.shouldShowReviewScreen();
     _isNetworkOn = _connectivityProvider.isNetworkOn;
     _syncNodeStateSubscription = _syncNodeStateStream.listen(_handleNodeSyncState);
     _walletBalance = _walletProvider
         .fetchWalletBalanceMap()
         .map((key, balance) => MapEntry(key, AnimatedBalanceData(balance.total, balance.total)));
     _walletProvider.walletLoadStateNotifier.addListener(updateWalletBalances);
+
+    _isBalanceHidden = _preferenceProvider.isBalanceHidden;
+    _fakeBalanceTotalAmount = _preferenceProvider.fakeBalanceTotalAmount;
+    _fakeBalanceMap = _preferenceProvider.getFakeBalanceMap();
   }
 
+  bool get isBalanceHidden => _isBalanceHidden;
+  bool get isReviewScreenVisible => _isReviewScreenVisible;
+  bool get isTermsShortcutVisible => _isTermsShortcutVisible;
   bool get shouldShowLoadingIndicator => !_isFirstLoaded && _nodeSyncState == NodeSyncState.syncing;
   List<WalletListItemBase> get walletItemList => _walletProvider.walletItemListNotifier.value;
   bool? get isNetworkOn => _isNetworkOn;
+  int? get fakeBalanceTotalAmount => _fakeBalanceTotalAmount;
+  Map<int, dynamic> get fakeBalanceMap => _fakeBalanceMap;
   Map<int, AnimatedBalanceData> get walletBalanceMap => _walletBalance;
 
   void _handleNodeSyncState(NodeSyncState syncState) {
@@ -62,6 +84,12 @@ class WalletListViewModel extends ChangeNotifier {
     }
   }
 
+  void hideTermsShortcut() {
+    _isTermsShortcutVisible = false;
+    _visibilityProvider.hideTermsShortcut();
+    notifyListeners();
+  }
+
   Future<void> updateWalletBalances() async {
     final updatedWalletBalance = _updateBalanceMap(_walletProvider.fetchWalletBalanceMap());
     _walletBalance = updatedWalletBalance;
@@ -81,6 +109,50 @@ class WalletListViewModel extends ChangeNotifier {
   void onWalletProviderUpdated(WalletProvider walletProvider) {
     _walletProvider = walletProvider;
     notifyListeners();
+  }
+
+  void onPreferenceProviderUpdated() {
+    /// 잔액 숨기기 변동 체크
+    if (_isBalanceHidden != _preferenceProvider.isBalanceHidden) {
+      _setIsBalanceHidden(_preferenceProvider.isBalanceHidden);
+    }
+
+    /// 가짜 잔액 총량 변동 체크 (on/off 판별)
+    if (_fakeBalanceTotalAmount != _preferenceProvider.fakeBalanceTotalAmount) {
+      _setFakeBlancTotalAmount(_preferenceProvider.fakeBalanceTotalAmount);
+      _setFakeBlanceMap(_preferenceProvider.getFakeBalanceMap());
+    }
+
+    /// 지갑별 가짜 잔액 변동 체크
+    if (_fakeBalanceMap != _preferenceProvider.getFakeBalanceMap()) {
+      // map이 변경되는 경우는 totalAmount가 변경되는 것과 관련이 없음
+      _setFakeBlanceMap(_preferenceProvider.getFakeBalanceMap());
+    }
+
+    notifyListeners();
+  }
+
+  void _setIsBalanceHidden(bool value) {
+    _isBalanceHidden = value;
+    notifyListeners();
+  }
+
+  void _setFakeBlancTotalAmount(int? value) {
+    _fakeBalanceTotalAmount = value;
+    notifyListeners();
+  }
+
+  void _setFakeBlanceMap(Map<int, dynamic> value) {
+    _fakeBalanceMap = value;
+    notifyListeners();
+  }
+
+  void updateAppReviewRequestCondition() async {
+    await AppReviewService.increaseAppRunningCountIfRejected();
+  }
+
+  int? getFakeBalance(int id) {
+    return _fakeBalanceMap[id] ?? _fakeBalanceTotalAmount;
   }
 
   void onNodeProviderUpdated() {
