+++ lib/widgets/card/wallet_item_card.dart
@@ -5,7 +5,6 @@ import 'package:coconut_wallet/localization/strings.g.dart';
 import 'package:coconut_wallet/model/wallet/balance.dart';
 import 'package:coconut_wallet/model/wallet/multisig_signer.dart';
 import 'package:coconut_wallet/utils/balance_format_util.dart';
-import 'package:coconut_wallet/utils/colors_util.dart';
 import 'package:coconut_wallet/widgets/animated_balance.dart';
 import 'package:coconut_wallet/widgets/icon/wallet_item_icon.dart';
 import 'package:flutter/material.dart';
@@ -21,7 +20,7 @@ class WalletItemCard extends StatelessWidget {
   final int colorIndex;
   final bool isLastItem;
   final bool isBalanceHidden;
-  final int? fakeBlance;
+  final int? fakeBalance;
   final List<MultisigSigner>? signers;
   final WalletImportSource walletImportSource;
   final BitcoinUnit currentUnit;
@@ -36,17 +35,17 @@ class WalletItemCard extends StatelessWidget {
     this.colorIndex = 0,
     required this.isLastItem,
     this.isBalanceHidden = false,
-    this.fakeBlance,
+    this.fakeBalance,
     this.signers,
     this.walletImportSource = WalletImportSource.coconutVault,
   });
 
   @override
   Widget build(BuildContext context) {
-    final displayFakeBalance = fakeBlance != null
+    final displayFakeBalance = fakeBalance != null
         ? currentUnit == BitcoinUnit.btc
-            ? satoshiToBitcoinString(fakeBlance!)
-            : addCommasToIntegerPart(fakeBlance!.toDouble())
+            ? satoshiToBitcoinString(fakeBalance!)
+            : addCommasToIntegerPart(fakeBalance!.toDouble())
         : '';
     // final isExternalWallet = walletImportSource != WalletImportSource.coconutVault;
     final row = ShrinkAnimationButton(
@@ -77,7 +76,7 @@ class WalletItemCard extends StatelessWidget {
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
                   isBalanceHidden
-                      ? fakeBlance != null
+                      ? fakeBalance != null
                           ? FittedBox(
                               fit: BoxFit.scaleDown,
                               alignment: Alignment.centerLeft,
